@@ -13,7 +13,6 @@ import os
 import sys
 from datetime import datetime
 from pathlib import Path
-from typing import Optional
 
 # Default log format
 DEFAULT_FORMAT = "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
@@ -64,8 +63,8 @@ class ColoredFormatter(logging.Formatter):
 
     def __init__(
         self,
-        fmt: Optional[str] = None,
-        datefmt: Optional[str] = None,
+        fmt: str | None = None,
+        datefmt: str | None = None,
         use_colors: bool = True,
     ):
         """
@@ -135,7 +134,7 @@ def get_log_level_from_env() -> int:
     return level_map.get(level_str, logging.INFO)
 
 
-def get_log_file_path() -> Optional[Path]:
+def get_log_file_path() -> Path | None:
     """
     Get the log file path from environment or default location.
 
@@ -157,9 +156,9 @@ def get_log_file_path() -> Optional[Path]:
 
 
 def setup_logging(
-    level: Optional[int] = None,
+    level: int | None = None,
     verbose: bool = False,
-    log_file: Optional[Path] = None,
+    log_file: Path | None = None,
     enable_file_logging: bool = True,
     use_colors: bool = True,
 ) -> logging.Logger:
@@ -332,7 +331,7 @@ def get_matching_log_path() -> Path:
 
 
 def setup_matching_logger(
-    log_file: Optional[Path] = None,
+    log_file: Path | None = None,
     level: int = logging.DEBUG,
 ) -> logging.Logger:
     """
